♻ Refactor: 댓글 및 답글 작성 Component #31
AuthorInfo.tsx 파일 위치 변경
불필요한 주석 삭제
