+++ src/main/reactfront/src/components/common/authorinfo/AuthorInfo.tsx
